Built constructor, config and cancel methods
The constructor initializes with several variables to be used by future methods. This library is just beginning and doesn't utilize every variable yet.
The config method takes an object that configures your callback variables to fit the circumstance.
Cancel does just that. If you cancel your callback, it won't execute when called.