Shiny new features
New "when" method: takes 1 or 2 arguments. Pass in a string to create a custom event and listen for it, or pass in two arguments, an object and the event to listen for.
New "if" method: pass a function to be evaluated before invoking your callback. If that function returns falsy, your callback will not be invoked and the falsy value will be returned.
New "invoke" method: this will test your conditional, and if passing, invoke your callback. If there are any errors, that error object will be passed to your error callback (although that callback isn't required).